requirements.txt
flask
google-cloud-aiplatform
google-cloud-storage
gunicorn
